Clean up test running
- Consolidate readline test code in test/utils.js
- Move responsibility for DEBUG_TESTS to run()
- Remove unused cwd option
